requirements.txt
# PyTorch with CUDA 12.1
torch --index-url https://download.pytorch.org/whl/cu121
torchvision --index-url https://download.pytorch.org/whl/cu121
torchaudio --index-url https://download.pytorch.org/whl/cu121

# Hugging Face tools
transformers
accelerate
datasets

# Whisper by OpenAI
openai-whisper

# OCR and image processing
PyMuPDF
pytesseract
pillow

# Text-to-speech
TTS

# Audio input/output
sounddevice
soundfile

